🚀 Create Release Workflow (#25)
* 🚀 Create Release Workflow

This github action makes use of original build tasks to create each of the artifacts and zips them up for easy consumption by the c# package

* 🔥 Delete unnecessary make install

Make install was failing and not completing. It is already installed on the system due to cmake. Removed the call for cleaner build logs.